feat: implement execute_index_commands and execute_transaction_commands
- execute_index_commands: fully implement CREATE INDEX and DROP INDEX
  - Use PropertyIndex.has_index, create_index, drop_index methods
  - Support OR REPLACE and IF NOT EXISTS for CREATE INDEX
  - Support IF EXISTS for DROP INDEX
  - Populate index with existing nodes when creating
  - Proper error handling for missing indexes/labels/properties

- execute_transaction_commands: implement BEGIN, COMMIT, ROLLBACK
  - Use TransactionManager.begin_write(), commit(), abort()
  - Flush storage on commit/rollback for durability
  - Note: full session context support would require storing transaction handles per session

- Remove all Future comments and mock implementations
- All functions now use real implementations, not placeholders
- Fix clippy warnings

+++ nexus-core/src/index/mod.rs
@@ -799,7 +799,7 @@ impl PropertyIndex {
             // Update stats: subtract entries from this index
             let mut removed_entries = 0u64;
             for bitmap in tree.values() {
-                removed_entries += bitmap.len() as u64;
+                removed_entries += bitmap.len();
             }
             stats.total_entries = stats.total_entries.saturating_sub(removed_entries);
             stats.indexed_properties = trees.len() as u32;
